chore: splitStringByComma 파일 오타 수정
